Rename `SearchableScope` to "SearchableMainContentScope"
